Add CI build using GitHub actions
Motivations:
the localproxy is cross-platform so I've added CI script that will build
the localproxy on OSX, Ubuntu and Windows each time a commit is pushed
that way we can use it as a sanity check.

Modifications:
Added the CI script, updated CMakeLists.txt for windows and updated the
windows build instructions.