fix #45 Allowing localproxy source port exposure via docker
Motivation:
To forward traffic from the host machine to the docker container, we
need to expose the listening port, and bind the localproxy listener to
0.0.0.0 address since "localhost" address is only accessible from within
the container.

Modification:
Updated the instructions to use the correct localproxy option to bind to
0.0.0.0 if the user needs to access localproxy from outside the
container, And updated the docker-run.sh script to accept
a port as an argument.

Test:
1. Ran the localproxy using docker, confirmed it's binding to 0.0.0.0
   and that I can SSH from outside the container (the host).
2. Ran the localproxy natively, confirmed it's binding to localhost and
   that I can SSH from the host.