Fix bus error from boost socket.shutdown in tcp_socket_error() and link statically to OpenSSL (#145)
* Fix bus error from boost socket.shutdown in tcp_socket_error()

Under MacOS (Darwin 23.1, arm64), building with Boost 1.81.0, localproxytest test-cases 'Test source mode" and 'Test source mode with client token' fail due to a bus error when calling connection->socket_.shutdown in tcp_adapter_proxy::tcp_socket_error. The exact same issue is triggered in localproxy, when a connection that's been setup through the tunnel, is subsequently closed (e.g. using SSH to login to the remote end and then logging out).

* Force static libs for OpenSSL

Update CMakeLists.txt to force linking with static versions of the OpenSSL libraries (libssl and libcrypto) when both static and share-object libs are available, in fashion similar to that used for Protobuf and Boost.